Create autoDATE
The add_expense function now accepts a date parameter that can be None. If None, the current date is automatically used,  datetime.today().strftime('%Y-%m-%d').
The user input for the date in the main function is made optional. If the user presses Enter without typing a date, None is passed to add_expense, triggering the use of the current date.